requirements.txt
astapi
uvicorn[standard]
kubernetes
jinja2
requests
python-dotenv
